+++ app/Console/Commands/BlockTypeContractsAuditCommand.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Console\Commands;
-
-use WebBlocks\Cms\Console\BlockTypeContractsAuditCommand as PackageBlockTypeContractsAuditCommand;
-
-class BlockTypeContractsAuditCommand extends PackageBlockTypeContractsAuditCommand {}