+++ app/Console/Commands/ImportDemoMedia.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Console\Commands;
-
-use WebBlocks\Cms\Console\ImportDemoMedia as PackageImportDemoMedia;
-
-class ImportDemoMedia extends PackageImportDemoMedia {}