+++ app/Console/Commands/ResetPrimitiveBlocksCommand.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Console\Commands;
-
-use WebBlocks\Cms\Console\ResetPrimitiveBlocksCommand as PackageResetPrimitiveBlocksCommand;
-
-class ResetPrimitiveBlocksCommand extends PackageResetPrimitiveBlocksCommand {}