+++ app/Console/Commands/SiteCloneCommand.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Console\Commands;
-
-use WebBlocks\Cms\Console\SiteCloneCommand as PackageSiteCloneCommand;
-
-class SiteCloneCommand extends PackageSiteCloneCommand {}