+++ app/Console/Commands/SiteDeleteCommand.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Console\Commands;
-
-use WebBlocks\Cms\Console\SiteDeleteCommand as PackageSiteDeleteCommand;
-
-class SiteDeleteCommand extends PackageSiteDeleteCommand {}