+++ app/Console/Commands/SyncCoreBlockTypesCommand.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Console\Commands;
-
-use WebBlocks\Cms\Console\SyncCoreBlockTypesCommand as PackageSyncCoreBlockTypesCommand;
-
-class SyncCoreBlockTypesCommand extends PackageSyncCoreBlockTypesCommand {}