+++ app/Models/BlockButtonTranslation.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace App\Models;
-
-class BlockButtonTranslation extends \WebBlocks\Cms\Models\BlockButtonTranslation {}