+++ app/Models/BlockContactFormTranslation.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace App\Models;
-
-class BlockContactFormTranslation extends \WebBlocks\Cms\Models\BlockContactFormTranslation {}