+++ app/Models/BlockGalleryItemTranslation.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace App\Models;
-
-class BlockGalleryItemTranslation extends \WebBlocks\Cms\Models\BlockGalleryItemTranslation {}