+++ app/Models/BlockImageTranslation.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace App\Models;
-
-class BlockImageTranslation extends \WebBlocks\Cms\Models\BlockImageTranslation {}